refactor: ai_report.py 리팩토링
